requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
timm>=0.9.0
einops>=0.6.0
tqdm>=4.65.0
Django>=4.2.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7 